+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 - UI: minor errors on menu entries tips.
 - Smart shuffle: foo_playcount was not being properly detected (instead looking for foo_enhanced_playcount).
 - Tags cache: crash due to wrong path at include code.
+- Tags: crash in some corner cases using WEIGHT method.
 
 ## [7.4.0] - 2024-07-30
 ### Added